refactor: modularize Electron main, fix preload paths, prune unused IPC
