+++ src/main/app/lifecycle.ts
@@ -1,0 +1,17 @@
+import { app, BrowserWindow } from 'electron'
+import { createMainWindow } from './window'
+
+export function registerAppLifecycle() {
+  app.on('window-all-closed', () => {
+    if (process.platform !== 'darwin') {
+      app.quit()
+    }
+  })
+
+  app.on('activate', () => {
+    if (BrowserWindow.getAllWindows().length === 0) {
+      createMainWindow()
+    }
+  })
+}
+
